move to Lagrange.Core.Message.Filter
