+++ Lagrange.Core/Message/Filter/MessageFilter.cs
@@ -1,8 +1,7 @@
-using System.Collections;
-using Lagrange.Core.Message.Entity;
-using Lagrange.Core.Message.FilterRule;
 
-namespace Lagrange.Core.Message;
+using Lagrange.Core.Message.Filter.Rule;
+
+namespace Lagrange.Core.Message.Filter;
 
 /// <summary>
 /// MessageFilter is a static class that provides a set of rules to filter the message chain.
@@ -12,12 +11,14 @@ internal static class MessageFilter
     /// <summary>
     /// The filter rules, result of the predicate is the index of the message entity that should be removed, -1 means the message entity should be kept.
     /// </summary>
-    private static readonly List<IMessageFilterRule> _rules = new();
+    private static readonly IMessageFilterRule[] _rules;
 
     static MessageFilter()
     {
-        _rules.Add(new ForwardRule());
-        _rules.Add(new ImageRule());
+        _rules = new IMessageFilterRule[] {
+            new ForwardTrailingAtAndTextFilterRule(),
+            new OldAndInvalidImageFilterRule(),
+        };
     }
 
     public static void Filter(MessageChain chain)
